Fix conversion for role binding to cluster role
This change makes it so that the conversion for authorization
policy correctly handles when a RBAC role binding references a
cluster role.  The code incorrectly assumed that the role binding's
namespace was the referenced role's namespace as well.  Now we use
the role's kind to determine its namespace.  This was missed by the
fuzzer tests because they only made role bindings to roles.  The
tests have been updated to create bindings to both types of roles.

Signed-off-by: Monis Khan <mkhan@redhat.com>